update: Echo, Gorm, MySQL Driver のアップデート

+++ src/domain/model/model.go
@@ -1,6 +1,6 @@
 package model
 
 type Aggregation struct {
-	Count   int
+	Count   int64
 	Average float32
 }
